digit_predictor_app/requirements.txt
streamlit
numpy
joblib
pillow
scikit-learn